Fix 23 code review issues across 7 severity categories
Critical fixes:
- CarryForward now respects credit-normal accounts (AP, revenue, equity)
- GL posting and sale completion wrapped in transactions (crash-safe)
- HMAC verification uses constant-time comparison, rejects default key
- Background thread DB access guarded with Debug.Assert

High fixes:
- double→decimal for all price/discount input parsing (IEEE-754 safety)
- FIFO cost: removed dead code, clamped negative stock to zero
- Sale+stock movements atomic in single transaction
- Price2/Price3 collision: skip tier price when qty-break is configured
- PrintReceiptAsync shows MessageBox on failure instead of silent swallow
- Sync column names validated with regex to prevent SQL injection

Medium/Low fixes:
- Pragmas re-applied on DB reconnect
- Discount display uses stored DiscValue instead of recalculating
- CounterRepository uses IClock instead of DateTime.Now
- SaleForm uses shared SaleRepository and IClock fields
- Zero-item sales rejected with guard clause
- UpdateService refuses unsigned update packages

+++ Kasir/Forms/Master/CreditCardForm.cs
@@ -83,14 +83,14 @@ namespace Kasir.Forms.Master
             {
                 if (dlg.ShowDialog(this) == DialogResult.OK)
                 {
-                    double fee;
-                    if (!double.TryParse(dlg.Values[2], out fee)) fee = 0;
+                    decimal fee;
+                    if (!decimal.TryParse(dlg.Values[2], out fee)) fee = 0;
 
                     var card = new CreditCard
                     {
                         CardCode = dlg.Values[0],
                         Name = dlg.Values[1],
-                        FeePct = (int)(fee * 100),
+                        FeePct = (int)(fee * 100m),
                         AccountCode = dlg.Values[3],
                         ChangedBy = _currentUserId
                     };
@@ -113,11 +113,11 @@ namespace Kasir.Forms.Master
             {
                 if (dlg.ShowDialog(this) == DialogResult.OK)
                 {
-                    double fee;
-                    if (!double.TryParse(dlg.Values[1], out fee)) fee = 0;
+                    decimal fee;
+                    if (!decimal.TryParse(dlg.Values[1], out fee)) fee = 0;
 
                     card.Name = dlg.Values[0];
-                    card.FeePct = (int)(fee * 100);
+                    card.FeePct = (int)(fee * 100m);
                     card.AccountCode = dlg.Values[2];
                     card.ChangedBy = _currentUserId;
